# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Express API with Swagger",
    "description": "A book CRUD API"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "tags": ["Test endpoints"],
        "summary": "Returns a hello world message",
        "responses": {
          "200": {
            "description": "Hello world message"
          }
        }
      }
    },
    "/otraruta": {
      "post": {
        "tags": ["Test endpoints"],
        "summary": "Returns a fixed JSON",
        "responses": {
          "200": {
            "description": "Fixed JSON"
          }
        }
      }
    },
    "/test-connection": {
      "get": {
        "tags": ["Test endpoints"],
        "summary": "Test the database connection",
        "responses": {
          "200": {
            "description": "Connection established or greeting returned."
          },
          "500": {
            "description": "Failed to establish connection or other server error."
          }
        }
      }
    },
    "/book": {
      "get": {
        "tags": ["Books"],
        "summary": "Get all books",
        "responses": {
          "200": {
            "description": "All book data"
          }
        }
      },
      "post": {
        "tags": ["Books"],
        "summary": "Add a book",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add book message"
          }
        }
      }
    },
    "/book/{id}": {
      "get": {
        "tags": ["Books"],
        "summary": "Get a specific book by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book details"
          },
          "404": {
            "description": "Book not found"
          }
        }
      },
      "put": {
        "tags": ["Books"],
        "summary": "Update a book",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "author": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated successfully"
          },
          "404": {
            "description": "Book not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["Books"],
        "summary": "Delete a specific book by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted Book"
          },
          "404": {
            "description": "Book not found"
          }
        }
      }
    }
  }
}